Move data-related classes into data package
